+++ src/form-elements/FormElementAutocomplete.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react'
+import { formService } from '@oneblink/apps'
 import { generateHeaders } from '@oneblink/apps/dist/services/fetch'
 
 import FormElementOptions from '../components/FormElementOptions'
@@ -150,7 +151,8 @@ const AutocompleteFetch = React.memo(function AutocompleteFetch({
         throw new Error(text)
       }
 
-      return await response.json()
+      const data = await response.json()
+      return formService.parseFormElementOptionsSet(data)
     },
     [searchUrl],
   )