+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Captured exceptions for certain element failures with Sentry
+
 ## [0.2.3] - 2021-04-15
 
 ### Changed