+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for `displayAsCurrency` prop in calculation element
 - `zoom` property to location form elements submission data
 
 ### Fixed